git-svn-id: svn+ssh://market.scs.stanford.edu/home/u1/asbestos/repo/jos64/trunk@330 927ebb28-5305-0410-b3ca-c4b640df69b7

+++ lib/lwip/jos64/arch/sys_arch.c
@@ -204,7 +204,7 @@ sys_thread_new(void (* thread)(void *arg), void *arg, int prio)
     threads[i].start = thread;
     threads[i].arg = arg;
 
-    int container = start_arg;
+    uint64_t container = start_arg;
     int r = thread_create(container, &sys_thread_entry, (void*)i, &threads[i].tobj);
     if (r < 0)
 	panic("lwip: sys_thread_new: cannot create: %s\n", e2s(r));
